Add some conditions so setting TARGET_FRAMEWORK to netcoreapp* works

+++ classes/dotnet.bbclass
@@ -1,7 +1,7 @@
 inherit dotnet-environment
 
 DEPENDS +=" mono-native"
-RDEPENDS_${PN}+=" mono libsystemnative"
+RDEPENDS_${PN}+=" mono libsystemnative libcurl libgssapi-krb5"
 
 BUILD="${WORKDIR}/build"
 PACKAGES = "${PN}-dbg ${PN}"
@@ -12,7 +12,7 @@ FILES_${PN}-dbg = "/opt/${PN}/*.pdb"
 # Some packages (EF for sqlite, Kestrel) have *.so files bundle.
 # This uses a sledge hammer aproach to fix the build
 PACKAGEBUILDPKGD_remove = "split_and_strip_files"
-INSANE_SKIP_${PN} = "file-rdeps arch"
+INSANE_SKIP_${PN} = "file-rdeps arch staticdev"
 
 TARGET_FRAMEWORK ?= "net471"
 MSBUILD_EXTRA_ARGS ?= ""
@@ -24,14 +24,18 @@ do_configure () {
 do_compile () {
     rm -rf ${BUILD}
     dotnet publish -c Release -o ${BUILD} -f ${TARGET_FRAMEWORK} ${MSBUILD_EXTRA_ARGS}
-    rm ${BUILD}/System.Diagnostics.Tracing.dll
+    if [$TARGET_FRAMEWORK = "net471"]; then
+        rm ${BUILD}/System.Diagnostics.Tracing.dll
+    fi
 }
 
 do_install () {
     install -d ${D}/opt/${PN}
     cp -dRf ${BUILD}/* ${D}/opt/${PN}
-    find ${D}/opt/${PN} -type d -exec chmod 755 {} +
-    find ${D}/opt/${PN} -type f -exec chmod 644 {} +
-    chmod 755 ${D}/opt/${PN}/*.exe
+    if [$TARGET_FRAMEWORK = "net471"]; then
+        find ${D}/opt/${PN} -type d -exec chmod 755 {} +
+        find ${D}/opt/${PN} -type f -exec chmod 644 {} +
+        chmod 755 ${D}/opt/${PN}/*.exe
+    fi
 }
